order modules and fix twitter bug
