+++ scripts/scripts.js
@@ -27,19 +27,7 @@ function buildHeroBlock(main) {
   }
 }
 
-function injectScript(src) {
-  window.scriptsLoaded = window.scriptsLoaded || [];
 
-  if (window.scriptsLoaded.indexOf(src)) {
-    const head = document.querySelector('head');
-    const script = document.createElement('script');
-
-    script.src = src;
-    script.setAttribute('async', 'true');
-    head.append(script);
-    window.scriptsLoaded.push(src);
-  }
-}
 
 function createEmbedWrap(a, vendor) {
   const div = document.createElement('div');
@@ -72,61 +60,7 @@ function preDecorateEmbed(main) {
   });
 }
 
-function createIframe(a, vendor) {
-  const div = a.nextElementSibling;
-  const embed = a.pathname;
-  const id = embed.split('/').pop();
-  let source;
-  let className;
-  let allow;
 
-  a.remove();
-
-  if (vendor === 'youtube') {
-    source = `https://www.youtube.com/embed/${id}`;
-    className = 'youtube-player';
-    allow = 'encrypted-media; accelerometer; gyroscope; picture-in-picture';
-  } else if (vendor === 'spotify') {
-    source = `https://open.spotify.com/embed/episode/${id}`;
-    className = 'spotify-player';
-    allow = 'autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture';
-  } else if (vendor === 'wistia') {
-    source = `https://fast.wistia.net/embed/iframe/${id}`;
-    className = 'wistia-player';
-    allow = 'autoplay; clipboard-write; encrypted-media; fullscreen;';
-  }
-
-  div.innerHTML = `<iframe src="${source}" 
-        class="${className}"
-        allowfullscreen  
-        allow="${allow}"
-        loading="lazy">
-    </iframe>`;
-}
-
-export function decorateEmbed() {
-  window.embedAnchors?.youTubeAnchors?.forEach((a) => {
-    createIframe(a, 'youtube');
-  });
-  window.embedAnchors?.spotifyAnchors?.forEach((a) => {
-    createIframe(a, 'spotify');
-  });
-  window.embedAnchors?.wistiaAnchors?.forEach((a) => {
-    createIframe(a, 'wistia');
-  });
-}
-
-export function decorateTwitterFeed(main) {
-  const anchors = main.getElementsByTagName('a');
-  const twitterAnchors = Array.from(anchors).filter((a) => a.href.includes('twitter'));
-
-  twitterAnchors.forEach((a) => {
-    a.innerText = `Tweets by ${a.pathname.split('/').pop()}`;
-    a.setAttribute('data-height', '500px');
-    a.classList.add('twitter-timeline');
-    injectScript('https://platform.twitter.com/widgets.js');
-  });
-}
 
 /**
  * Builds all synthetic blocks in a container element.
